+++ src/platform/mod.rs
@@ -924,18 +924,22 @@ mod platform_impl {
 
 #[cfg(all(
     target_arch = "wasm32",
-    target_os = "unknown",
+    any(target_os = "emscripten", target_os = "unknown"),
     feature = "wasm-bindgen"
 ))]
 mod platform_impl {
-    #[cfg(all(feature = "audioworklet", target_feature = "atomics"))]
+    #[cfg(all(
+        target_os = "unknown",
+        feature = "audioworklet",
+        target_feature = "atomics"
+    ))]
     use crate::host::audioworklet::Host as AudioWorkletHost;
     use crate::host::webaudio::Host as WebAudioHost;
     use crate::traits::HostTrait as _;
 
     impl_platform_host!(
         WebAudio => WebAudioHost,
-        #[cfg(all(feature = "audioworklet", target_feature = "atomics"))] AudioWorklet => AudioWorkletHost,
+        #[cfg(all(target_os = "unknown", feature = "audioworklet", target_feature = "atomics"))] AudioWorklet => AudioWorkletHost,
         #[cfg(feature = "custom")] Custom => super::CustomHost
     );
 
@@ -1004,7 +1008,7 @@ mod platform_impl {
     target_os = "android",
     all(
         target_arch = "wasm32",
-        target_os = "unknown",
+        any(target_os = "emscripten", target_os = "unknown"),
         feature = "wasm-bindgen"
     ),
 )))]